+++ clientele/utils.py
@@ -1,3 +1,4 @@
+import os
 import re
 
 from openapi_core import Spec
@@ -117,3 +118,13 @@ def union_for_py_ver(union_items: list) -> str:
         return " | ".join(union_items)
     else:
         return f"typing.Union[{', '.join(union_items)}]"
+
+
+def get_client_project_directory_path(output_dir: str) -> str:
+    """
+    Returns a dot-notation path for the client directory.
+    Assumes that the `clientele` command is being run in the
+    project root directory.
+    """
+    dot_notation = ".".join(os.path.join(output_dir).split("/")[:-1])
+    return dot_notation