+++ tests/test_client/client.py
@@ -2,10 +2,7 @@ from __future__ import annotations
 
 import typing  # noqa
 
-from . import (
-    http,  # noqa
-    schemas,  # noqa
-)
+from tests.test_client import http, schemas  # noqa
 
 
 def complex_model_request_complex_model_request_get() -> schemas.ComplexModelResponse: